init commit changing log name to time log
