+++ app/DomainValueObjects/TimeLog/TimePuncher.php
@@ -1,0 +1,29 @@
+<?php 
+declare (strict_types = 1);
+namespace App\DomainValueObjects\TimeLog;
+
+use App\DomainValueObjects\UUIDGenerator\UUID;
+use App\DomainValueObjects\TimeLog\TimeStamp\TimeStamp;
+use App\DomainValueObjects\Location\Location;
+
+
+//Exceptions
+
+
+
+class TimePuncher
+{
+    private $timeStamp =  null;
+    private $location = null;
+
+    public function __construct(UUID $uuid, TimeStamp $timeStamp, Location $location)
+    {
+        $this->uuid = $uuid;
+        $this->timeStamp = $timeStamp;
+        $this->location = $location;
+        $this->validation();
+    }
+
+    private function validation()
+    { }
+}
