--- app/Http/Controllers/Api/Log/ClockInDomain/Contracts/ClockInContract.php
@@ -1,7 +1,0 @@
-<?php
-namespace App\Http\Controllers\Api\Log\ClockInDomain\Contracts;
-
-interface ClockInContract
-{
-    public function clockIn(string $currentLocation, string $timeStamp);
-}
