--- app/Http/Controllers/Api/Log/ClockOutDomain/Contracts/ClockOutContract.php
@@ -1,7 +1,0 @@
-<?php
-namespace App\Http\Controllers\Api\Log\ClockOutDomain\Contracts;
-
-interface ClockOutContract
-{
-    public function clockOut(string $currentLocation, string $timeStamp, string $logUuid);
-}
