--- app/Http/Controllers/Api/Log/TImePuncher/Contracts/TimePuncherContract.php
@@ -1,8 +1,0 @@
-<?php
-namespace App\Http\Controllers\Api\Log\TimePuncher\Contracts;
-
-interface TimePuncherContract
-{
-    public function getUserLocationAndUserTimeSheetId($user, $currentLocation): array;
-    public function getUserLocation($user, $currentLocation);
-}
