+++ app/Http/Controllers/Api/TimeLog/ClockInDomain/Contracts/ClockInContract.php
@@ -1,0 +1,7 @@
+<?php
+namespace App\Http\Controllers\Api\TimeLog\ClockInDomain\Contracts;
+
+interface ClockInContract
+{
+    public function clockIn(string $currentLocation, string $timeStamp);
+}
