+++ app/Http/Controllers/Api/TimeLog/ClockOutDomain/Contracts/ClockOutContract.php
@@ -1,0 +1,7 @@
+<?php
+namespace App\Http\Controllers\Api\TimeLog\ClockOutDomain\Contracts;
+
+interface ClockOutContract
+{
+    public function clockOut(string $currentLocation, string $timeStamp, string $logUuid);
+}
