+++ app/Http/Controllers/Api/TimeLog/TImePuncher/Contracts/TimePuncherContract.php
@@ -1,0 +1,8 @@
+<?php
+namespace App\Http\Controllers\Api\TimeLog\TimePuncher\Contracts;
+
+interface TimePuncherContract
+{
+    public function getUserLocationAndUserTimeSheetId($user, $currentLocation): array;
+    public function getUserLocation($user, $currentLocation);
+}
