+++ app/Providers/ApiProviders/TimeLog/ClockOutServiceProvider.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Providers\ApiProviders\TimeLog;
+
+use Illuminate\Support\ServiceProvider;
+
+class ClockOutServiceProvider extends ServiceProvider
+{
+    /**
+     * Bootstrap services.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        //
+    }
+
+    /**
+     * Register services.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        $this->app->bind(
+            'App\Http\Controllers\Api\TimeLog\ClockOutDomain\Contracts\ClockOutContract',
+            'App\Http\Controllers\Api\TimeLog\ClockOutDomain\Services\ClockOutService'
+        );
+    }
+}
