+++ app/TimeLog.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class TimeLog extends Model
+{
+    protected $fillable = ['id', 'user_id', 'time_sheet_id', 'clock_in', 'clock_out', 'log_change_stack'];
+    public $incrementing = false;
+}
